Initial Controller Push
Controllers created for both Users and Recipes, as well as some additions to Bundle.properties for error messages. UsersController showing errors because Users.java does not have Long ID value, Users table needs ID column of Long values.

+++ recipesite/src/main/java/com/alexandjoe/recipesite/entity/Recipes.java
@@ -35,7 +35,7 @@ public class Recipes implements Serializable {
     @Id
     @Basic(optional = false)
     @Column(name = "ID")
-    private Integer id;
+    private Long id; // Changed for Controllers Line 22
     @Basic(optional = false)
     @Column(name = "NAME")
     private String name;
